test: login 테스트 다이나믹 테스트로 변경
