--- src/test/java/roomescape/auth/ui/LoginControllerTest.java
@@ -1,72 +1,0 @@
-package roomescape.auth.ui;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import io.restassured.RestAssured;
-import java.util.Map;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import roomescape.auth.dto.MemberResponse;
-import roomescape.auth.dto.TokenRequest;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class LoginControllerTest {
-
-    private static final String EMAIL = "email@email.com";
-    private static final String PASSWORD = "1234";
-    private static final String NAME = "백호";
-
-    @LocalServerPort
-    int port;
-
-    @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
-    }
-
-    @DisplayName("로그아웃을 한다.")
-    @Test
-    void tokenLogin() {
-        Map<String, String> params = Map.of("email", EMAIL, "password", PASSWORD, "name", NAME);
-
-        RestAssured
-                .given().log().all()
-                .body(params)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/members")
-                .then().log().all()
-                .statusCode(HttpStatus.CREATED.value());
-
-        String accessToken = RestAssured
-                .given().log().all()
-                .body(new TokenRequest(EMAIL, PASSWORD))
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/login")
-                .then().log().all().extract().header("Set-Cookie").split("=")[1];
-
-        MemberResponse member = RestAssured
-                .given().log().all()
-                .cookie("token", accessToken)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/login/check")
-                .then().log().all()
-                .statusCode(HttpStatus.OK.value()).extract().as(MemberResponse.class);
-
-        assertThat(member.name()).isEqualTo(NAME);
-
-        RestAssured
-                .given().log().all()
-                .cookie("token", accessToken)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/logout")
-                .then().log().all()
-                .statusCode(HttpStatus.OK.value());
-    }
-}
